requirements_fastapi.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.31
psycopg2-binary==2.9.9

# Google AI
google-genai==1.2.0

# Other dependencies
python-dotenv==1.0.1
python-multipart==0.0.6
